+++ Settings.py
@@ -1,0 +1,16 @@
+
+
+def show_options():
+    opts = """
+    Welcome To CareerSoft Bank
+    1. Deposit
+    2. Withdraw
+    3. Balance
+    4. Mini Statement
+    5. Exit
+    """
+    print(opts)
+    ch = input("Please select option from above: ")
+
+    return ch
+
